pyproject.toml
[tool.poetry]
name = "seedwebbase"
version = "0.1.0"
description = ""
authors = ["Tom Camp <tomcamp@protonmail.ch>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
